Cleaned up structure
